GCM Handling updated
