+++ PushNotification/PushNotification/PushNotification.Plugin.Abstractions/IPushNotificationListener.cs
@@ -7,13 +7,35 @@ using System.Threading.Tasks;
 
 namespace PushNotification.Plugin
 {
+    /// <summary>
+    /// Push Events Listener
+    /// </summary>
     public interface IPushNotificationListener
     {
+        /// <summary>
+        /// On Message Received
+        /// </summary>
+        /// <param name="Parameters"></param>
+        /// <param name="deviceType"></param>
         void OnMessage(IDictionary<string, object> Parameters,DeviceType deviceType);
+        /// <summary>
+        /// On Registered
+        /// </summary>
+        /// <param name="Token"></param>
+        /// <param name="deviceType"></param>
         void OnRegistered(string Token, DeviceType deviceType);
+        /// <summary>
+        /// On Unregistered
+        /// </summary>
+        /// <param name="deviceType"></param>
         void OnUnregistered( DeviceType deviceType);
+        /// <summary>
+        /// OnError
+        /// </summary>
         void OnError(string message,DeviceType deviceType);
-
+        /// <summary>
+        /// Should Show Notification
+        /// </summary>
         bool ShouldShowNotification();
     }
 }
