+++ PushNotification/PushNotification/PushNotification.Plugin.Abstractions/PushNotificationKeys.cs
@@ -6,22 +6,75 @@ using System.Threading.Tasks;
 
 namespace PushNotification.Plugin.Abstractions
 {
+    /// <summary>
+    /// Push Notification Keys
+    /// </summary>
     public static class  PushNotificationKey
     {
+
+      /// <summary>
+      /// Type
+      /// </summary>
       public const string Type = "type";
+      /// <summary>
+      /// Error
+      /// </summary>
       public const string Error = "error";
+      /// <summary>
+      /// Domain Name
+      /// </summary>
 	  public const string DomainName = "CrossPushNotification";
+      /// <summary>
+      /// Title
+      /// </summary>
       public const string Title = "title";
+      /// <summary>
+      /// Text
+      /// </summary>
       public const string Text = "text";
+      /// <summary>
+      /// Subtitle
+      /// </summary>
       public const string Subtitle = "subtitle";
+      /// <summary>
+      /// Message
+      /// </summary>
 	  public const string Message = "message";
+      /// <summary>
+      /// Silent
+      /// </summary>
       public const string Silent = "silent";
+      /// <summary>
+      /// Token
+      /// </summary>
 	  public const string Token = "token";
+      /// <summary>
+      /// App Version
+      /// </summary>
 	  public const string AppVersion = "appVersion";
+      /// <summary>
+      /// IntentFromGcmMessage
+      /// </summary>
       public const string IntentFromGcmMessage = "com.google.android.c2dm.intent.RECEIVE";
+      /// <summary>
+      /// BackOffMilliseconds
+      /// </summary>
       public const string BackOffMilliseconds = "backoff_ms";
+      /// <summary>
+      /// ErrorServiceNotAvailable
+      /// </summary>
       public const string ErrorServiceNotAvailable= "SERVICE_NOT_AVAILABLE";
+      /// <summary>
+      /// Tag
+      /// </summary>
       public const string Tag = "tag";
+      /// <summary>
+      /// Id
+      /// </summary>
       public const string Id = "id";
+      /// <summary>
+      /// RegistrationComplete
+      /// </summary>
+      public const string RegistrationComplete = "registrationComplete";
     }
 }
