+++ PushNotification/PushNotification/PushNotification.Plugin.Abstractions/PushNotificationNotInitializedException.cs
@@ -6,12 +6,20 @@ using System.Threading.Tasks;
 
 namespace PushNotification.Plugin.Abstractions
 {
+    /// <summary>
+    /// Push Notification not Initialized Exception class
+    /// </summary>
     public class PushNotificationNotInitializedException : Exception
     {
+          /// <summary>
+          /// Default Contructor
+          /// </summary>
           public PushNotificationNotInitializedException()
           {
           }
-
+          /// <summary>
+          /// Constructor with message
+          /// </summary>
           public PushNotificationNotInitializedException(string message): base(message)
           {
           }
