+++ PushNotification/PushNotification/PushNotification.Plugin.Android/PushNotificationsReceiver.cs
@@ -9,6 +9,7 @@ using Android.OS;
 using Android.Runtime;
 using Android.Views;
 using Android.Widget;
+using Android.Gms.Gcm;
 
 [assembly: Permission(Name = "@PACKAGE_NAME@.permission.C2D_MESSAGE")]
 [assembly: UsesPermission(Name = "android.permission.WAKE_LOCK")]
@@ -18,16 +19,14 @@ using Android.Widget;
 
 namespace PushNotification.Plugin
 {
-    [BroadcastReceiver(Permission = "com.google.android.c2dm.permission.SEND")]
+    [BroadcastReceiver(Permission = "com.google.android.c2dm.permission.SEND", Exported=true)]
     [IntentFilter(new string[] { "com.google.android.c2dm.intent.RECEIVE" }, Categories = new string[] { "@PACKAGE_NAME@" })]
-    [IntentFilter(new string[] { "com.google.android.c2dm.intent.REGISTRATION" }, Categories = new string[] { "@PACKAGE_NAME@" })]
-    [IntentFilter(new string[] { "com.google.android.gcm.intent.RETRY" }, Categories = new string[] { "@PACKAGE_NAME@" })]
-    public class PushNotificationsReceiver : WakefulBroadcastReceiver
+    public class PushNotificationsReceiver : GcmReceiver
     {
-        public override void OnReceive(Context context, Intent intent)
+        /*public override void OnReceive(Context context, Intent intent)
         {
           
-			var serviceIntent = new Intent(context,typeof(PushNotificationImplementation));
+			var serviceIntent = new Intent(context,typeof(PushNotificationGcmListener));
             serviceIntent.AddFlags(ActivityFlags.IncludeStoppedPackages);
             serviceIntent.ReplaceExtras(intent.Extras);
             serviceIntent.SetAction( intent.Action);
@@ -35,6 +34,6 @@ namespace PushNotification.Plugin
           
 
             ResultCode = Result.Ok;
-        }
+        }*/
     }
 }
