+++ PushNotification/PushNotification/PushNotification.Plugin.Android/packages.config
@@ -1,7 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Xamarin.Android.Support.v4" version="21.0.3.0" targetFramework="MonoAndroid41" />
-  <package id="Xamarin.Android.Support.v7.AppCompat" version="21.0.3.0" targetFramework="MonoAndroid41" />
-  <package id="Xamarin.Android.Support.v7.MediaRouter" version="21.0.3.0" targetFramework="MonoAndroid41" />
-  <package id="Xamarin.GooglePlayServices" version="22.0.0.2" targetFramework="MonoAndroid41" />
+  <package id="Xamarin.Android.Support.v4" version="22.2.0.0" targetFramework="MonoAndroid50" />
+  <package id="Xamarin.GooglePlayServices.Base" version="25.0.0.0" targetFramework="MonoAndroid50" />
+  <package id="Xamarin.GooglePlayServices.Gcm" version="25.0.0.0" targetFramework="MonoAndroid50" />
 </packages>
