+++ PushNotification/PushNotification/PushNotification.Plugin/CrossPushNotification.cs
@@ -62,6 +62,11 @@ namespace PushNotification.Plugin
     {
       get
       {
+          //Should always initialize plugin before use
+        if (!CrossPushNotification.IsInitialized)
+        {
+              throw NewPushNotificationNotInitializedException();
+        }
         var ret = Implementation.Value;
         if (ret == null)
         {
@@ -83,6 +88,13 @@ namespace PushNotification.Plugin
     internal static Exception NotImplementedInReferenceAssembly()
     {
       return new NotImplementedException("This functionality is not implemented in the portable version of this assembly.  You should reference the NuGet package from your main application project in order to reference the platform-specific implementation.");
+    }
+
+    internal static PushNotificationNotInitializedException NewPushNotificationNotInitializedException()
+    {
+        string description = "CrossPushNotification Plugin is not initialized. Should initialize before use with CrossPushNotification Initialize method. Example:  CrossPushNotification.Initialize<CrossPushNotificationListener>()";
+
+        return new PushNotificationNotInitializedException(description);
     }
   }
 }
